Desarrollo del addIncidente del IncidenteService

+++ src/main/java/com/NatanielSanchez/SistemaReporteIncidentes/controllers/requestDTOs/DetalleProblemaRequestDTO.java
@@ -12,5 +12,5 @@ public class DetalleProblemaRequestDTO
 {
     @NotEmpty
     private long id_problema;
-    private long[] estimaciones;
+    private int[] estimaciones;
 }
